+++ R/LearnerS4XGBoost.R
@@ -541,7 +541,9 @@ setMethod("..train", signature(object="familiarXGBoost", data="dataObject"),
             # Check if the model trained at all.
             if(!is.null(object@messages$error)) return(callNextMethod(object=object))
             
-            # Add model
+            # Add model. Note that we do not use xgboost::xgb.save.raw to save
+            # raw model data, since this prevents the model being used for
+            # variable importance.
             object@model <- model
             
             # Add the contrast references to model_list
@@ -574,6 +576,11 @@ setMethod("..predict", signature(object="familiarXGBoost", data="dataObject"),
               # Check if the data is empty.
               if(is_empty(data)) return(callNextMethod())
               
+              # Load model through unserialisation.
+              if(inherits(object@model, "raw")){
+                object@model <- xgboost::xgb.load.raw(object@model)
+              }
+              
               # Encode data so that the features are the same as in the training.
               encoded_data <- encode_categorical_variables(data=data,
                                                            object=object,
@@ -582,7 +589,7 @@ setMethod("..predict", signature(object="familiarXGBoost", data="dataObject"),
               
               # Get an empty prediction table.
               prediction_table <- get_placeholder_prediction_table(object=object,
-                                                                   data=encoded_data$encoded_dat,
+                                                                   data=encoded_data$encoded_data,
                                                                    type=type)
               
               # Make predictions. If the booster object is DART type, predict()
@@ -733,6 +740,12 @@ setMethod("..vimp", signature(object="familiarXGBoost"),
             # Check that required packages are loaded and installed.
             require_package(object, "vimp")
             
+            # Load model through unserialisation. Note that this currently does
+            # not produce a model that can be used to infer variable importance.
+            if(inherits(object@model, "raw")){
+              object@model <- xgboost::xgb.load.raw(object@model)
+            }
+            
             # Use xgboost::xgb.importance function from xgboost to extract a
             # data.table.
             xgboost_score <- tryCatch(xgboost::xgb.importance(model=object@model),
@@ -857,18 +870,18 @@ setMethod("..set_recalibration_model", signature(object="familiarXGBoost", data=
 setMethod(".trim_model", signature(object="familiarXGBoost"),
           function(object, ...){
             
+            # Set is_trimmed to TRUE.
+            object@is_trimmed <- TRUE
+            
+            # Prevent trimming of raw, serialised model information.
+            if(inherits(object@model, "raw")) return(object)
+            
             # Update model by removing the call.
             object@model$call <- call("trimmed")
             
             # Add show.
             object <- .capture_show(object)
-            
-            # Remove raw
-            object@model$raw <- NULL
-            
-            # Set is_trimmed to TRUE.
-            object@is_trimmed <- TRUE
-            
+
             # Default method for models that lack a more specific method.
             return(object)
           })