+++ NEWS.md
@@ -1,3 +1,5 @@
+# Development
+
 # Version 1.1.1 (Flamboyant Flamingo)
 
 ## Minor changes